refactor(agent): separate selected SWM transfer ownership
